backend/requirements.txt
Django==5.0.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
celery[redis]==5.4.0
django-celery-beat==2.6.0
psycopg2-binary==2.9.9
redis==5.1.1
requests==2.32.3
weasyprint==62.3
openpyxl==3.1.5
python-decouple==3.8
